+++ astrea/astro/astro/propagation/equations_of_motion/instances/EquinoctialVop.cpp
@@ -11,7 +11,7 @@
  * have received a copy of the GNU General Public License along with Astrea. If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include <astro/propagation/equations_of_motion/EquinoctialVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/EquinoctialVop.hpp>
 
 #include <mp-units/math.h>
 #include <mp-units/systems/angular/math.h>
