+++ astrea/astro/astro/propagation/equations_of_motion/instances/EquinoctialVop.test.cpp
@@ -17,22 +17,23 @@
 #include <units/units.hpp>
 
 #include <astro/platforms/Vehicle.hpp>
-#include <astro/propagation/equations_of_motion/CowellsMethod.hpp>
+#include <astro/propagation/equations_of_motion/instances/EquinoctialVop.hpp>
 #include <astro/propagation/force_models/ForceModel.hpp>
 #include <astro/state/State.hpp>
-#include <astro/state/orbital_elements/instances/Cartesian.hpp>
+#include <astro/state/orbital_elements/instances/Equinoctial.hpp>
 #include <astro/systems/AstrodynamicsSystem.hpp>
 #include <tests/utilities/comparisons.hpp>
 
-using namespace astrea;
-using namespace astro;
-using namespace mp_units;
+using mp_units::angular::unit_symbols::rad;
 using mp_units::si::unit_symbols::km;
 using mp_units::si::unit_symbols::s;
 
-class CowellsMethodTest : public testing::Test {
+using namespace astrea;
+using namespace astro;
+
+class EquinoctialTest : public testing::Test {
   public:
-    CowellsMethodTest() :
+    EquinoctialTest() :
         eom(forces)
     {
     }
@@ -46,7 +47,7 @@ class CowellsMethodTest : public testing::Test {
     AstrodynamicsSystem sys;
     Date epoch;
     ForceModel forces;
-    CowellsMethod eom;
+    EquinoctialVop eom;
 };
 
 
@@ -57,33 +58,18 @@ int main(int argc, char** argv)
 }
 
 
-TEST_F(CowellsMethodTest, GetExpectedSet)
+TEST_F(EquinoctialTest, GetExpectedSet)
 {
-    ASSERT_EQ(eom.get_expected_set_id(), OrbitalElements::get_set_id<Cartesian>());
+    ASSERT_EQ(eom.get_expected_set_id(), OrbitalElements::get_set_id<Equinoctial>());
 }
 
-TEST_F(CowellsMethodTest, Derivative)
+TEST_F(EquinoctialTest, Derivative)
 {
-    Cartesian cart0           = Cartesian::LEO(sys.get_mu());
-    CartesianPartial expected = CartesianPartial(
-        cart0.get_vx(), cart0.get_vy(), cart0.get_vz(), -0.0081347028957142863 * km / (s * s), 0.0 * km / (s * s), 0.0 * km / (s * s)
-    );
-    State state0(cart0, epoch, sys);
+    Equinoctial equi0 = Equinoctial::LEO(sys.get_mu());
+    EquinoctialPartial expected =
+        EquinoctialPartial(0.0 * km / s, 0.0 * 1 / s, 0.0 * 1 / s, 0.0 * 1 / s, 0.0 * 1 / s, 0.0010780076129942077 * rad / s);
+    State state(equi0, epoch, sys);
 
-    OrbitalElementPartials dstate = eom.compute_dynamics(state0, sat, noForce, noForce);
-    ASSERT_EQ_ORB_PART(expected, dstate, REL_TOL);
-}
-
-// Vallado, Ex. 8.5
-TEST_F(CowellsMethodTest, DerivativeValladoEx85)
-{
-    Cartesian cart0{ -605.790796 * km,   -5870.230422 * km,  3493.051916 * km,
-                     -1.568251 * km / s, -3.702348 * km / s, -6.479485 * km / s };
-    CartesianPartial expected = CartesianPartial(
-        cart0.get_vx(), cart0.get_vy(), cart0.get_vz(), 0.00074873079 * km / (s * s), 0.00725534667 * km / (s * s), -0.00431725847 * km / (s * s)
-    );
-    State state0(cart0, epoch, sys);
-
-    OrbitalElementPartials dstate = eom.compute_dynamics(state0, sat, noForce, noForce);
+    OrbitalElementPartials dstate = eom.compute_dynamics(state, sat, noForce, noForce);
     ASSERT_EQ_ORB_PART(expected, dstate, REL_TOL);
 }
