+++ astrea/astro/astro/propagation/equations_of_motion/instances/J2MeanVop.test.cpp
@@ -17,10 +17,10 @@
 #include <units/units.hpp>
 
 #include <astro/platforms/Vehicle.hpp>
-#include <astro/propagation/equations_of_motion/EquinoctialVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/J2MeanVop.hpp>
 #include <astro/propagation/force_models/ForceModel.hpp>
 #include <astro/state/State.hpp>
-#include <astro/state/orbital_elements/instances/Equinoctial.hpp>
+#include <astro/state/orbital_elements/instances/Keplerian.hpp>
 #include <astro/systems/AstrodynamicsSystem.hpp>
 #include <tests/utilities/comparisons.hpp>
 
@@ -31,12 +31,9 @@ using mp_units::si::unit_symbols::s;
 using namespace astrea;
 using namespace astro;
 
-class EquinoctialTest : public testing::Test {
+class J2MeanTest : public testing::Test {
   public:
-    EquinoctialTest() :
-        eom(forces)
-    {
-    }
+    J2MeanTest() {}
 
     void SetUp() override {}
 
@@ -46,8 +43,7 @@ class EquinoctialTest : public testing::Test {
     Vehicle sat;
     AstrodynamicsSystem sys;
     Date epoch;
-    ForceModel forces;
-    EquinoctialVop eom;
+    J2MeanVop eom;
 };
 
 
@@ -58,17 +54,15 @@ int main(int argc, char** argv)
 }
 
 
-TEST_F(EquinoctialTest, GetExpectedSet)
-{
-    ASSERT_EQ(eom.get_expected_set_id(), OrbitalElements::get_set_id<Equinoctial>());
-}
+TEST_F(J2MeanTest, GetExpectedSet) { ASSERT_EQ(eom.get_expected_set_id(), OrbitalElements::get_set_id<Keplerian>()); }
 
-TEST_F(EquinoctialTest, Derivative)
+TEST_F(J2MeanTest, Derivative)
 {
-    Equinoctial equi0 = Equinoctial::LEO(sys.get_mu());
-    EquinoctialPartial expected =
-        EquinoctialPartial(0.0 * km / s, 0.0 * 1 / s, 0.0 * 1 / s, 0.0 * 1 / s, 0.0 * 1 / s, 0.0010780076129942077 * rad / s);
-    State state(equi0, epoch, sys);
+    Keplerian kep0 = Keplerian::LEO();
+    KeplerianPartial expected =
+        KeplerianPartial(0.0 * km / s, 0.0 * 1 / s, 0.0 * rad / s, 0.0 * rad / s, 0.0 * rad / s, 0.0010780076129942077 * rad / s);
+
+    State state(kep0, epoch, sys);
 
     OrbitalElementPartials dstate = eom.compute_dynamics(state, sat, noForce, noForce);
     ASSERT_EQ_ORB_PART(expected, dstate, REL_TOL);
