+++ astrea/astro/astro/propagation/equations_of_motion/instances/KeplerianVop.cpp
@@ -11,7 +11,7 @@
  * have received a copy of the GNU General Public License along with Astrea. If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include <astro/propagation/equations_of_motion/KeplerianVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/KeplerianVop.hpp>
 
 #include <iostream>
