+++ astrea/astro/tests/nasa_6dof_checkcases/orbital/driver.cpp
@@ -29,14 +29,14 @@
 #include <astro/frames/frames.hpp>
 #include <astro/platforms/vehicles/Spacecraft.hpp>
 
-#include <astro/propagation/equations_of_motion/CowellsMethod.hpp>
-#include <astro/propagation/equations_of_motion/EquinoctialVop.hpp>
-#include <astro/propagation/equations_of_motion/J2MeanVop.hpp>
-#include <astro/propagation/equations_of_motion/KeplerianVop.hpp>
-#include <astro/propagation/equations_of_motion/TwoBody.hpp>
+#include <astro/propagation/equations_of_motion/instances/CowellsMethod.hpp>
+#include <astro/propagation/equations_of_motion/instances/EquinoctialVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/J2MeanVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/KeplerianVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/TwoBody.hpp>
 
-#include <astro/propagation/force_models/instances/AtmosphericForce.hpp>
 #include <astro/propagation/force_models/ForceModel.hpp>
+#include <astro/propagation/force_models/instances/AtmosphericForce.hpp>
 #include <astro/propagation/force_models/instances/NBodyForce.hpp>
 #include <astro/propagation/force_models/instances/OblatenessForce.hpp>
 #include <astro/propagation/force_models/instances/SolarRadiationPressure.hpp>
