+++ astrea/astro/tests/propagation/equinoctial-vop/driver.cpp
@@ -17,7 +17,7 @@
 #include <units/units.hpp>
 
 #include <astro/platforms/vehicles/Spacecraft.hpp>
-#include <astro/propagation/equations_of_motion/EquinoctialVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/EquinoctialVop.hpp>
 #include <astro/propagation/force_models/ForceModel.hpp>
 #include <astro/propagation/numerical/Integrator.hpp>
 #include <astro/state/orbital_elements/OrbitalElements.hpp>
