+++ astrea/astro/tests/propagation/keplerian-vop/driver.cpp
@@ -17,7 +17,7 @@
 #include <units/units.hpp>
 
 #include <astro/platforms/vehicles/Spacecraft.hpp>
-#include <astro/propagation/equations_of_motion/KeplerianVop.hpp>
+#include <astro/propagation/equations_of_motion/instances/KeplerianVop.hpp>
 #include <astro/propagation/force_models/ForceModel.hpp>
 #include <astro/propagation/numerical/Integrator.hpp>
 #include <astro/state/orbital_elements/OrbitalElements.hpp>
