+++ astrea/astro/tests/propagation/thruster-scheduling/driver.cpp
@@ -18,7 +18,7 @@
 
 #include <astro/platforms/thrusters/Thruster.hpp>
 #include <astro/platforms/vehicles/Spacecraft.hpp>
-#include <astro/propagation/equations_of_motion/TwoBody.hpp>
+#include <astro/propagation/equations_of_motion/instances/TwoBody.hpp>
 #include <astro/propagation/event_detection/Event.hpp>
 #include <astro/propagation/event_detection/Schedule.hpp>
 #include <astro/propagation/event_detection/events/TurnThrustersOff.hpp>
